refactor(ws): extract device inbound dispatch to shared service
Move per-message dispatch logic from cms/routers/ws.py into a new

cms/services/device_inbound.py module. Pure mechanical refactor — zero

behaviour change. All existing tests pass unchanged.

This is preparation for issue #344 stage 2b.2, where a WebPubSub

webhook receiver will reuse the same dispatch_device_message() function

to handle inbound device frames that arrive via Azure Web PubSub rather

than directly over WebSocket.

- New: cms/services/device_inbound.py (449 lines)

  - InboundContext dataclass for per-connection mutable state

  - dispatch_device_message(msg, ctx, db, send) — single-message handler

  - rotate_api_key(device, db, send) — replaces WS-coupled helper

  - _resolve_asset_for_device moved here verbatim

- Modified: cms/routers/ws.py (-375 lines, +38 lines)

  - Replaces 270-line dispatch loop with dispatch_device_message call

  - send callable wraps websocket.send_json

  - Handshake/register/disconnect lifecycle stays in ws.py

  - Pruned imports that moved to device_inbound
